+++ NEWS.md
@@ -1,7 +1,14 @@
+# gaussplotR 0.2.0
+
+* Added a `fit_gaussian_2D()` function to apply any of several available methods
+to fit a 2D Gaussian to supplied data
+* `predict_gaussian_2D()` has been enhanced to allow any method to be used in
+`fit_gaussian_2D()` and then fed into the predict function
+
 # gaussplotR 0.1.6
 
 * Added a `compute_gaussian_volume()` function to compute the volume under a 
-given 2D gaussian
+given 2D Gaussian
 
 # gaussplotR 0.1.4